Fix manual hazard-tracking.
Problem was putting trust in our documentation which stated fence updates are scheduled at the end of the encoder. This is not true. The GPU driver *may* do so, but it may also encode them straight away, which is what happens. Public documentation is wrong, Metal header file has the correct information.