[Hdx,Maya] Move Render Tag ownership from the RenderSetupTask to RenderTask.
This change is primarily to support the Maya scene delegate, but has a nice side effect of simplifying the RenderTask's render tag support
and removes the complex handshacking between setup task and render task.

Previously, the render setup task would be provided the render tags through it's Params structure.  RenderSetupTask would provide the tags
to the Sync algorithm and then communicate them through the task context to the RenderTask for execute.  This was additionaly complicated
by sometimes the RenderTask owned the setup task.

Now, full ownership of render tags is within the RenderTask.  It uses the new RenderTag dirty bit and scene delegate API to obtain them.

As the Maya scene delegate uses one RenderTask per shape, this now means it no longer needs to generate a union of render tags for the setup task.

Due to some dependencies, it was also necessary to convert the Pick task to the new API at the same time.

Changes:
- The pick and render tasks now has seperate ownership of render tags, through a _renderTags attribute.
- When syncing these task, the _renderTags attribute is populated through the new _GetTaskRenderTags() API when HdChangeTracker::DirtyRenderTags is set.
- The task's GetRenderTags() method, now just returns the value of the attribute.
- HdxRenderTask::_GetRenderTagsForExecute() is no longer needed and Execute can just use GetRenderTags() as the values are the same.
- Removed render tags from the HdxRenderTaskParams and HdxPickTaskParams structures.
- Removed _renderTags attribute and getter from HdxRenderSetupTask.
- Updated task controller to not set render tags in params structures.
- Task controller's SetRenderTags() method now sets the render tags parameter for the pick task as well as the render tasks.
- Hdx's renderTags token, which was used for inter-task communication is no longer needed.
- Hdx_UnitTestDelegate and Maya scene delegates now implements the GetTaskRenderTags() scene delegate API.
- UsdImagingGLEngine no longer needs to set render tags in the render params structure.
- UsdImagingGLEngine::TestIntersection() wasn't setting the render tags on the task controller.
- Maya scene delegate's GetPickingTasks() method now takes a collection and uses this to set the render tags for the picking task.
- Maya scene delegate's GetRenderTasks() method now sets the render tags on a per task basis rather than creating a union from all collections.

(Internal change: 1973936)