trace: fix stale pointer after appending event containers
TraceEventContainer::Append failed to update the _nextEvent pointer so
that subsequent calls to emplace_back could stomp on the appended
events or write to invalid memory.  This does not occur in the system
today because Append is only used to finalize a collection; no further
calls to emplace_back are made on the appended-to container.

(Internal change: 1978661)