In wrapSchemaBase, replace the __getattribute__ method as was already being done in wrapObject. This overloaded __getattribute__ actually allows any of the UsdSchemaBase APIs to be called (as all of the methods at this level can still return useful information). But its purpose is to block calls to APIs defined on subclasses.
Provided similar protection for UsdGeomPrimvar, which is an odd case that is
like a schema class, but operates on a single attribute and so is not derived
from SchemaBase.

Added some test cases, including an explicit test of issue #872, a crash when
calling UsdGeom.Xformable().ClearXformOpOrder(). TestUsdRiSplineAPI had to be
updated because a test in there was not expecting the RuntimeException from
an invalid object.