In the old HdxIntersector (deleted from repo in changelist 1971956) we hardcoded alphaThreshold to 0.0 inside HdxIntersector::params. This 0.0 alpha threshold value was then passed to RenderPassState.
An alpha threshold of 0.0 disabled the 'discard' we have in our glslfx function ShouldDiscardByAlpha().

In the refactor of PickTask, we pass alpha threshold from TaskController to RenderPassState.
In UsdView, the params passed into TaskController come via _MakeHydraUsdImagingGLRenderParams() which has in it this logic:

        params.alphaThreshold =
            renderParams.enableSampleAlphaToCoverage ? 0.1f : 0.5f;

Since alpha_to_coverage is disabled by PickTask, UsdImagingGL provides us a alphaThreshold of 0.5.
This means most transparent pixels are discarded during picking.

Seting 0.0 for alpha threshold in _MakeHydraUsdImagingGLRenderParams would solve the issue.
We decided however that we want to start reducing our alphaThreshold usage, so I'm removing it completely from PickTask and hardcoding it to 0.0

(Internal change: 1986681)