First changelist to add volumetric rendering to hydra's OpenGL backend. It will render volume prims as constant density ellipsoids with size given by the prim's extent if the environment variable HDST_ENABLE_EXPERIMENTAL_VOLUME_ELLIPSOID_STANDINS is set).
The changelist adds an rprim HdStVolume which draws the bounding box of the prim using triangles and uses a shader key to run a simple volume shader running in the OIT render pass to integrate along the camera rays and rays to the light source. There is a new GeometricShader::PrimitiveType PRIM_VOLUME to support this.

For the new test case, a new argument (-camera) was added to testUsdImagingGLBasicDrawing.

(Internal change: 1986942)
(Internal change: 1987044)