Exposed number of completed samples from hdEmbree as render stats.
Use the new facility for render delegates to expose render stats in
hdEmbree.  For now, usdview hard-codes the stats keys that it knows
about.  In the future, we may need to relax and make it more general.

Portions of this change include portions merged from PR#888 thanks
to marktucker.

(Internal change: 1987352)