Change semantics of GetConnectedSource() to return false if the targeted output does not exist.  Previously, we misleadingly only required the host target *prim* to exist to indicate success.
UsdImagingMaterialAdapter was also not processing connections
to material inputs properly: if the material input has no value, then
we must take the value authored on the connected input, if it
exists.

Also update dox for affected methods, and add a new section to
UsdShade's overview.dox that explains connection semantics,
also fixing upstream/downstream confusion in the examples.

(Internal change: 1987950)