Fix buffer version hash to include shader param BAR
During batch validation (HdSt_DrawBatch::Validate), we compute the front draw item's bufferArraysHash and compare it with the batch's previous hash.
This hash incorporates the version number of all the BARs known to the draw item (the BARs themselves are "owned" by the RprimSharedData object on the Rprim).
HdDrawItem::GetBufferArraysHash factors the various primvar frequency BARs, topology BARs and instance BARs to compute the hash.
Living in the base class, it didn't factor in the shader param BAR managed by HdStSurfaceShader in HdStDrawItem, and thus meant that the batches and/or the dispatch buffer for the (indirect) draw batch was not updated correctly. This situation can be triggered via garbage collection of a BAR that shared the buffer array for the shader param data.
When garbage collection happens, the version number of the HdBufferArray is bumped up. Calling GetVersion on the HdBufferArrayRange returns the version of the underlying buffer array.

Changes:
- Add virtual _GetBufferArraysHash to HdDrawItem to allow derived classes to compute hashes of any buffers they manage/know about.
- Add implementation in HdStDrawItem that returns the version of the shader param BAR.

(Internal change: 1990382)