Fix memory leak for draw items
HdRepr holds a vector of (raw) pointers to draw items, and is responsible for deleting them during destruction (which happens on Rprim removal). Failure to do so leaks CPU memory.
While HdDrawItem doesn't manage any BAR (and instead delegates to RprimSharedData), HdStDrawItem holds a handle to the surface shader, which manages a BAR for the params array, leading to GPU memory that was never being released when using Storm (HdSt).

(Internal change: 1990426)