Making UsdImagingGLHydraMaterialAdapter follow outputs:volume of a material if outputs:surface is not connected. If it follows the outputs:volume, it uses the volumeShader technique instead of the surfaceShader technique from a GLSLFX file.
Ideally, the query of the render delegate to the scene delegate for a shader
would specify whether it is for a surface or volume shader. In this change, this
distinction is not made and the shader is used according to whether it is
consumed by a volume prim or not.
The proper time to fix this is when we switch from the
UsdImagingGLHydraMaterialAdapter to HdNetworkMaterialMap which will have
different entries for the surface and volume shader.

(Internal change: 1993410)