USD is now relocatable and can be built without having to install dependent libraries on the system
Details:
A script is run at the end of the build process to change the absolute paths inside the binaries to relative paths.
Resolved dependency on PySide, QT and OpenGL by copying the binaries to the folder
No longer downloading zlib, boost, TBB, GLEW, OpenEXR and OpenSubdiv. This is needed for building on Barli

Former-commit-id: be7310fe7f2d75bff45ee4b04ca15e17e40f05d7