[hd, hdSt] Improve the BAR interface for querying offsets and fix copy computation for interleaved resources.
When querying offsets in a BAR, we may care about the byte offset at which the BAR begins for a resource, or just the element offset in the underlying buffer array. The existing GetOffset API returned the element offset for VBO BARs and byte offset for interleaved UBO/SSBO BARs, and did not robustly handle migrating data via the GPU copy computation for interleaved BARs.

- Remove GetIndex()
- Replace GetOffset() with GetElementOffset() and GetByteOffset(resource). Fix clients to call the appropriate function.
- Update copyComputation to handle fixed size arrays (by factoring the tupleType's count member).

(Internal change: 2039755)