Add HgiShaderFunction and HgiShaderProgram. Add GL implementations for both.
Notes:

HgiGLConversions::GetShaderStages returns a vector of shader stages, because HgiShaderStage is a bitmask.
This is not immediately obvious here, but the fact it is a bitmask becomes very useful later on when we need
to define in what shader stages resource bindings are available etc.

HgiShaderStage does not currently include tessellation or geometry stages, because we do not yet need
them and these stages requier a bit more discussion since they are different between rendering api's.

(Internal change: 2040727)
(Internal change: 2040778)